Moves the tests to a dedicated subdir
None of the tests in the main file touch on implementation internals. By moving them to a separate file it makes it easier to manage.
